+++ Source/ios-framework/CodeGenTests/ToolTestProject52.swift
@@ -3,11 +3,19 @@
 //  ToolTestProject
 //
 //  Created by Uli Kusterer on 07.12.18.
-//  Copyright © 2018 Uli Kusterer. All rights reserved.
+//  Copyright 2018-2025 ObjectBox. All rights reserved.
 //
 
-import ObjectBox
+/// Tests the model JSON file is not overwritten on a generator ID error.
+///
+/// The existing model file (`expected/model/model52.before.json`) was generated
+/// using `ToolTestProject52_original.swift`. To assert there are no changes, it
+/// is also used as the expected model file (`expected/model/model52.json`).
+///
+/// The entity triggers an ID error (DuplicatePropertyID)by assigning the
+/// existing UID of `finalStop` to `lineName`.
 
+import ObjectBox
 
 class BusRoute: Entity {
     var id: EntityId<BusRoute> = 0